Add Authorization Checks

+++ Backend/controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator/check');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
 	const currentPage = req.query.page || 1;
@@ -47,18 +48,28 @@ exports.createPost = (req, res, next) => {
 	const imageUrl = req.file.path;
 	const title = req.body.title;
 	const content = req.body.content;
+	let creator;
 	const post = new Post({
 		title: title,
 		content: content,
 		imageUrl: imageUrl,
-		creator: { name: 'Maximilian' },
+		creator: { userId: req.userId, name: req.name },
 	});
 	post
 		.save()
 		.then((result) => {
+			return User.findById(req.userId);
+		})
+		.then((user) => {
+			creator = user;
+			user.posts.push(post);
+			return user.save();
+		})
+		.then((result) => {
 			res.status(201).json({
 				message: 'Post created successfully!',
-				post: result,
+				post: post,
+				creator: { _id: creator._id, name: creator.name },
 			});
 		})
 		.catch((err) => {
@@ -114,6 +125,11 @@ exports.updatePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized!');
+				error.statusCode = 403;
+				throw error;
+			}
 			if (imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl);
 			}
@@ -142,12 +158,23 @@ exports.deletePost = (req, res, next) => {
 				error.statusCode = 404;
 				throw error;
 			}
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized!');
+				error.statusCode = 403;
+				throw error;
+			}
 			// Check logged in user
 			clearImage(post.imageUrl);
 			return Post.findByIdAndRemove(postId);
 		})
 		.then((result) => {
-			console.log(result);
+			return User.findById(req.userId);
+		})
+		.then((user) => {
+			user.posts.pull(postId);
+			return user.save();
+		})
+		.then((result) => {
 			res.status(200).json({ message: 'Deleted post.' });
 		})
 		.catch((err) => {
